finished design and started on build and test (code)
I started on the foundation of the GUI using tkinter. At first I encountered a problem because my code couldn't run, but then I realised I forgot a root.mainloop() which was supposed to run the code. The next step would be to add a bit more for the UI like a label, or text box.